+++ src/main/java/race/domain/Car.java
@@ -1,4 +1,4 @@
-package race;
+package race.domain;
 
 public class Car {
     private final CarName carName;
