+++ src/main/java/race/domain/CarResult.java
@@ -1,4 +1,4 @@
-package race;
+package race.domain;
 
 public class CarResult {
     private String carName;
