+++ src/main/java/race/domain/CarsFactory.java
@@ -1,4 +1,4 @@
-package race;
+package race.domain;
 
 import java.util.Arrays;
 import java.util.List;
