+++ src/main/java/race/domain/MoveStrategy.java
@@ -1,4 +1,4 @@
-package race;
+package race.domain;
 
 public interface MoveStrategy {
     boolean canMove();
