+++ src/main/java/race/domain/Race.java
@@ -1,4 +1,4 @@
-package race;
+package race.domain;
 
 import java.util.List;
 import java.util.stream.Collectors;
