+++ src/main/java/race/domain/RandomMoveStrategy.java
@@ -1,4 +1,4 @@
-package race;
+package race.domain;
 
 import java.util.Random;
