+++ src/test/java/race/TestHelper.java
@@ -1,5 +1,7 @@
 package race;
 
+import race.domain.MoveStrategy;
+
 public class TestHelper {
     public static class AlwaysMoveStrategy implements MoveStrategy {
         @Override
