+++ src/test/java/race/domain/CarTest.java
@@ -1,10 +1,14 @@
-package race;
+package race.domain;
 
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
 import org.junit.jupiter.params.provider.ValueSource;
+import race.TestHelper;
+import race.domain.Car;
+import race.domain.CarName;
+import race.domain.MoveStrategy;
 
 import java.util.stream.Stream;
