+++ src/test/java/race/domain/CarsTest.java
@@ -1,9 +1,14 @@
-package race;
+package race.domain;
 
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
+import race.TestHelper;
+import race.domain.Car;
+import race.domain.CarName;
+import race.domain.Cars;
+import race.domain.MoveStrategy;
 
 import java.util.List;
 import java.util.stream.Collectors;
