+++ src/test/java/race/domain/RaceTest.java
@@ -1,9 +1,11 @@
-package race;
+package race.domain;
 
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
+import race.TestHelper;
+import race.domain.*;
 
 import java.util.List;
 import java.util.stream.Collectors;
