+++ src/test/java/race/domain/RandomMoveStrategyTest.java
@@ -1,7 +1,8 @@
-package race;
+package race.domain;
 
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
+import race.domain.RandomMoveStrategy;
 
 import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
